refactor(alignment): 拆出锚点抽取与结构识别，semantic_alignment 2808→1659
领域层四大模块此前完全不受行数门禁约束，长到 1800~2900 行。本轮先把四个
模块按当前行数封顶止血，再拆 semantic_alignment。

原计划两层切（锚点抽取器外移）不可行：_excluded_anchor_indices 反向依赖
目录/注释识别，闭包分析显示锚点块 756 行需额外拖入 410 行，其中 14 项仍被
留下的代码使用。那 410 行是"哪些段不是正文"的判定，锚点抽取与正文区域推断
共用，属底层而非锚点私有实现。故切为三层单向依赖：

  alignment_structure  482 行  标题行/目录页/编号注释块识别（无包内依赖）
  alignment_anchors    864 行  4 类锚点抽取器 + AnchorExtractorRegistry
  semantic_alignment  1659 行  嵌入、正文区域、DP、注释覆盖

分区经 AST 分层方向检查验证，无注入机制、无签名改动；代码按顶层节点整块
搬移，未改写任何函数体。HeadingAnchor 等仍从 semantic_alignment 重新导出，
text_alignment 与其余测试的调用面零变化。

行数上限只降不升：semantic_alignment 2825→1675，新增 alignment_anchors 875
与 alignment_structure 500。

全量 unittest discover 1981 项通过（1 跳过），对齐相关 105 项与拆分前基线
逐项一致；导入环守卫绿，顶层导入图仍为 0 环。
